+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/ExoPlayerImpl.java
@@ -372,6 +372,7 @@ import java.util.concurrent.TimeoutException;
               builder.livePlaybackSpeedControl,
               builder.releaseTimeoutMs,
               pauseAtEndOfMediaItems,
+              builder.dynamicSchedulingEnabled,
               applicationLooper,
               clock,
               playbackInfoUpdateListener,